requirements.txt
python==3.13.2
streamlit==1.43.2
google-generativeai==0.8.5